+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG: cfbd_json_py
 
+## 0.2.1 The "College Football is BACK! (2024)" Update
+- Fixed a bug in `cfbd_json_py.drives.get_cfbd_drives_info()` where the `ncaa_division` parameter would become malformed, the API call wouldn't filter by `ncaa_division`, but the API call would still be accepted.
+- Updated the package to comply with changes made in version `4.6.0` of the CFBD V1 API. These changes in this API version removed the `game_id` parameter from `cfbd_json_py.betting.get_cfbd_betting_lines()` and from `cfbd_json_py.games.get_cfbd_weather_info()`.
+- Updated the package version to `0.2.1`.
+
 ## 0.2.0 The "Patreon" Update
 - Re-implemented the process of storing a user's API key. If you have used `cfbd_json_py.utls.set_cfbd_api_token()` in the past, you do not need to do anything to migrate your API key to this new process.
 - The following functions require a user to subscribe to the [CFBD Patreon](https://www.patreon.com/collegefootballdata):